app-support: pubsub service changed to uint8_t

+++ app-support/pub-sub.c
@@ -13,7 +13,7 @@
 
 // the struct to keep each topic subscribed
 typedef struct topic {
-  uint16_t service;
+  uint8_t service;
   name_component_t identifier[2];
   uint32_t interval; // the time interval between two Interests
   uint32_t next_interest; // the time to send next Interest
@@ -64,13 +64,13 @@ _on_subscription_interest()
 }
 
 void
-ps_subscribe_to(uint16_t service, char* identifier, uint32_t identifier_len,
+ps_subscribe_to(uint8_t service, char* identifier, uint32_t identifier_len,
                 uint32_t frequency, ndn_on_content_published callback);
 
 void
-ps_publish_content(uint16_t service, uint16_t datatype, uint32_t datatype_len,
+ps_publish_content(uint8_t service, uint16_t datatype, uint32_t datatype_len,
                    uint8_t* content, uint32_t content_len);
 
 void
-ps_publish_command(uint16_t service, uint16_t action, char* identifier, uint32_t identifier_len,
+ps_publish_command(uint8_t service, uint16_t action, char* identifier, uint32_t identifier_len,
                    uint8_t* content, uint32_t content_len);
